app/requirements/production.txt
django==4.1.6
django-crispy-forms==1.14.0
django-allauth==0.52.0
django-ninja==0.20.0
gunicorn==20.1.0
tzdata==2022.7
psycopg2-binary==2.9.5
django-filter==22.1
django-pwa==1.1.0
django-request==1.6.2
crispy-bootstrap5==0.7
markdown==3.4.1
pillow==9.4.0
celery==5.2.7
flower==1.2.0
WeasyPrint==58.0
django-bootstrap-v5==1.0.11